updated pgm

+++ firmware/Adafruit_GFX_AS.cpp
@@ -53,10 +53,9 @@ POSSIBILITY OF SUCH DAMAGE.
   #include "Font7s.h"
 #endif
 
-
+// just two to start with
 #define pgm_read_byte(addr) (*(const unsigned char *)(addr))
-#define pgm_read_word(addr) (*(const uint8_t *)(addr))
-
+#define pgm_read_word(addr) (*(const unsigned short *)(addr))
 
 Adafruit_GFX_AS::Adafruit_GFX_AS(int16_t w, int16_t h):
   WIDTH(w), HEIGHT(h)
